# Don't record annotations other than payload event for gRPC